+++ packages/atom.io/__tests__/experimental/realtime-ipc/game-instance.bun.ts
@@ -2,9 +2,9 @@ import type { LoggerIcon, TokenDenomination } from "atom.io"
 import { AtomIOLogger, setState } from "atom.io"
 import { editRelations } from "atom.io/data"
 import { findState } from "atom.io/ephemeral"
+import { IMPLICIT } from "atom.io/internal"
 import * as RTS from "atom.io/realtime-server"
 
-import { IMPLICIT } from "../../../internal/src/store"
 import { gameContinuity, letterAtoms } from "./game-store"
 
 const parentSocket = new RTS.ParentSocket()
