+++ packages/atom.io/internal/src/families/dispose-from-store.ts
@@ -1,19 +1,23 @@
 import type { MoleculeToken, ReadableToken } from "atom.io"
-import { disposeMolecule } from "atom.io"
-import * as Internal from "atom.io/internal"
+
+import { disposeAtom } from "../atom"
+import { disposeMolecule } from "../molecule/dispose-molecule"
+import { disposeSelector } from "../selector"
+import type { Store } from "../store"
+import { IMPLICIT } from "../store"
 
 export function disposeFromStore(
 	token: MoleculeToken<any> | ReadableToken<any>,
-	store: Internal.Store = Internal.IMPLICIT.STORE,
+	store: Store = IMPLICIT.STORE,
 ): void {
 	switch (token.type) {
 		case `atom`:
 		case `mutable_atom`:
-			Internal.disposeAtom(token, store)
+			disposeAtom(token, store)
 			break
 		case `selector`:
 		case `readonly_selector`:
-			Internal.disposeSelector(token, store)
+			disposeSelector(token, store)
 			break
 		case `molecule`:
 			disposeMolecule(token, store)
