+++ packages/atom.io/internal/src/molecule/create-molecule-family.ts
@@ -1,0 +1,30 @@
+import type {
+	MoleculeConstructor,
+	MoleculeCreation,
+	MoleculeFamily,
+	MoleculeFamilyOptions,
+	MoleculeFamilyToken,
+} from "atom.io"
+
+import type { Store } from "../store"
+import { Subject } from "../subject"
+
+export function createMoleculeFamily<M extends MoleculeConstructor>(
+	options: MoleculeFamilyOptions<M>,
+	store: Store,
+): MoleculeFamilyToken<M> {
+	const subject = new Subject<MoleculeCreation<M>>()
+
+	const token = {
+		type: `molecule_family`,
+		key: options.key,
+		dependsOn: options.dependsOn ?? `all`,
+	} as const satisfies MoleculeFamilyToken<M>
+	const family = {
+		...token,
+		subject,
+		new: options.new,
+	} satisfies MoleculeFamily<M>
+	store.moleculeFamilies.set(options.key, family)
+	return token
+}
