+++ packages/atom.io/internal/src/molecule/index.ts
@@ -1,1 +1,5 @@
-export * from "./molecule"
+export * from "./create-molecule-family"
+export * from "./dispose-molecule"
+export * from "./grow-molecule-in-store"
+export * from "./make-molecule-in-store"
+export * from "./molecule-internal"
